fix: reindent only pre elements
