+++ src/core/examples.js
@@ -6,9 +6,9 @@
 // When an example is found, it is reported using the "example" event. This can
 // be used by a containing shell to extract all examples.
 
-import { addId, reindent } from "./utils";
+import { addId } from "./utils";
 import css from "text!../../assets/examples.css";
-import hyperHTML from "hyperhtml";
+import html from "hyperhtml";
 import { pub } from "./pubsubhub";
 
 export const name = "core/examples";
@@ -17,10 +17,10 @@ function makeTitle(conf, elem, num, report) {
   report.title = elem.title;
   if (report.title) elem.removeAttribute("title");
   const number = num > 0 ? ` ${num}` : "";
-  return hyperHTML`
+  return html`
   <div class="marker"><a class="self-link">${conf.l10n.example}${number}</a>${
     report.title
-      ? hyperHTML`<span class="example-title">: ${report.title}</span>`
+      ? html`<span class="example-title">: ${report.title}</span>`
       : ""
   }</div>`;
 }
@@ -33,7 +33,7 @@ export function run(conf) {
   if (!examples.length) return;
 
   document.head.insertBefore(
-    hyperHTML`<style>${css}</style>`,
+    html`<style>${css}</style>`,
     document.querySelector("link")
   );
 
@@ -58,24 +58,19 @@ export function run(conf) {
       const { id } = example;
       const selfLink = div.querySelector("a.self-link");
       selfLink.href = `#${id}`;
-      const children = Array.from(example.children);
-      children.forEach(child => {
-        child.innerHTML = reindent(child.innerHTML);
-      });
       pub("example", report);
     } else {
       const inAside = !!example.closest("aside");
       if (!inAside) ++number;
 
-      const reindentedHtml = reindent(example.innerHTML);
-      example.innerHTML = report.content = reindentedHtml;
+      report.content = example.innerHTML;
 
       // wrap
       example.classList.remove("example", "illegal-example");
       // relocate the id to the div
       const id = example.id ? example.id : null;
       if (id) example.removeAttribute("id");
-      const div = hyperHTML`
+      const div = html`
         <div class='example' id="${id}">
           ${makeTitle(conf, example, inAside ? 0 : number, report)}
           ${example.cloneNode(true)}
