+++ src/core/utils.js
@@ -265,25 +265,6 @@ export function normalizePadding(text = "") {
   return result;
 }
 
-/**
- * Removes common indents across the IDL texts,
- * so that indentation inside <pre> won't affect the rendered result.
- * @param {string} text IDL text
- */
-export function reindent(text) {
-  if (!text) {
-    return text;
-  }
-  // TODO: use trimEnd when Edge supports it
-  const lines = text.trimRight().split("\n");
-  while (lines.length && !lines[0].trim()) {
-    lines.shift();
-  }
-  const indents = lines.filter(s => s.trim()).map(s => s.search(/[^\s]/));
-  const leastIndent = Math.min(...indents);
-  return lines.map(s => s.slice(leastIndent)).join("\n");
-}
-
 // RESPEC STUFF
 export function removeReSpec(doc) {
   doc.querySelectorAll(".remove, script[data-requiremodule]").forEach(elem => {
